requirements.txt
Pillow
numpy
torch
numba